requirements.txt
Flask==1.0.2
alembic==1.0.9
Flask_SQLAlchemy==2.3.2
argon2_cffi==19.1.0
SQLAlchemy==1.3.3
toastedmarshmallow==0.2.6
PyJWT==1.7.1
marshmallow==2.19.2
